Move repository file and split it
