+++ backend/src/modules/_base/repository.ts
@@ -1,0 +1,4 @@
+export * from './db-repository';
+export * from './kv-repository';
+export * from './search-repository';
+export * from './common';
